requirements.txt
Django==2.2.*
psycopg2-binary>=2.8
djangorestframework==3.12.*
djangorestframework-simplejwt==4.6.*
djoser==2.1.*
django-cors-headers==3.7.*
django-filter==2.4.*
django-heroku==0.3.1
dj-database-url==0.5.0
django-nose==1.4.7
gunicorn==20.0.4
black==21.5b2
flake8==3.9.2
opencv-python==4.5.1.48
numba==0.55.1
numpy==1.20.1
mysqlclient 
django-cloudinary-storage
Pillow
django-environ
cloudinary